format code and minor refactor
